refactor(room): split RoomScreen monolithic styles into per-component files
Extract 4 sub-style creators from styles.ts (573 -> ~140 barrel):
- actionButton.styles.ts: action button + danger variant + message
- boardInfo.styles.ts: board info card + night progress indicator
- statusPanels.styles.ts: bottom panel + connection bar + banner
- modalMenu.styles.ts: host menu dropdown + seat confirm modal

Barrel composes via spread. No visual/behavioral change.
